# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: LedgerMeadow API
  version: 1.0.0
  description: Go-owned authenticated financial API. Monetary amounts are decimal strings in minor units.
servers:
  - url: http://127.0.0.1:8080
security:
  - clerkBearer: []
paths:
  /api/v1/bank-connections/link-token:
    post:
      operationId: createLinkToken
      responses:
        "200":
          description: Plaid Link token created.
          content: {application/json: {schema: {$ref: "#/components/schemas/LinkTokenResponse"}}}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "502": {$ref: "#/components/responses/ProviderUnavailable"}
...
components:
  securitySchemes:
    clerkBearer: {type: http, scheme: bearer, bearerFormat: JWT}
...
  responses:
...
    Unauthenticated:
      description: Valid Clerk session required.
      content: {application/json: {schema: {$ref: "#/components/schemas/ErrorResponse"}}}
    ProviderUnavailable:
      description: Plaid could not complete the operation.
      content: {application/json: {schema: {$ref: "#/components/schemas/ErrorResponse"}}}
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties: {code: {type: string}, message: {type: string}}
...
    LinkTokenResponse:
      type: object
      required: [link_token, expiration]
      properties: {link_token: {type: string}, expiration: {type: string, format: date-time}}
...